+++ integration/Makefile
@@ -1,3 +1,3 @@
 .PHONY: test
 test:
-	cargo test
+	cargo test --test supervisor
