+++ integration/features/balancer.feature
@@ -1,15 +1,15 @@
-# Feature: Balancer
+Feature: Balancer
 
-#     Scenario: Balancer can loadbalance
-#       Given balancer-1 is running at 0.0.0.0:8070, 0.0.0.0:8071 and reports metrics to 0.0.0.0:9125 every 1 second in balancer feature
-#       Given llamacpp-1 is running at 8080 with 4 slots in balancer feature
-#       Given agent-1 is running and observing llamacpp-1 in 127.0.0.1:8080, and registered at balancer-1 in 127.0.0.1:8070 in balancer feature
-#       Given llamacpp-2 is running at 8081 with 3 slots in balancer feature
-#       Given agent-2 is running and observing llamacpp-2 in 127.0.0.1:8081, and registered at balancer-1 in 127.0.0.1:8070 in balancer feature
+    Scenario: Balancer can loadbalance
+      Given balancer-1 is running at 0.0.0.0:8070, 0.0.0.0:8071 and reports metrics to 0.0.0.0:9125 every 1 second
+      Given llamacpp-1 is running at 8080 with 4 slots
+      Given agent-1 is running and observing llamacpp-1 in 127.0.0.1:8080, and registered at balancer-1 in 127.0.0.1:8070
+      Given llamacpp-2 is running at 8081 with 3 slots
+      Given agent-2 is running and observing llamacpp-2 in 127.0.0.1:8081, and registered at balancer-1 in 127.0.0.1:8070
 
-#       When 1 request is proxied to balancer-1 in 127.0.0.1:8071 in balancer feature
-#       Then balancer-1 must tell 1 slot is busy and 6 slots are idle in 127.0.0.1:8070 from agent-1 and agent-2 in balancer feature
-#       Then balancer-1 must return a successful response in 127.0.0.1:8071 in balancer feature
+      When 1 request is proxied to balancer-1 in 127.0.0.1:8071
+      Then balancer-1 must tell 1 slot is busy and 6 slots are idle in 127.0.0.1:8070 from agent-1 and agent-2
+      Then balancer-1 must return a successful response in 127.0.0.1:8071
 
 #     Scenario: Balancer cannot loadbalance
 #       Given balancer-1 is running at 0.0.0.0:8070, 0.0.0.0:8071 and reports metrics to 0.0.0.0:9125 every 1 second in balancer feature
