+++ integration/src/main.rs
@@ -1,5 +1,3 @@
-mod tests;
-
 fn main() {
     panic!("Use `cargo test` instead!")
 }
