--- integration/src/tests/mod.rs
@@ -1,19 +1,0 @@
-pub mod agent;
-pub mod balancer;
-pub mod supervisor;
-pub mod utils;
-
-#[cfg(test)]
-mod tests {
-    // use super::agent::tests::run_agent_tests;
-    // use super::balancer::tests::run_balancer_tests;
-    use super::supervisor::run_supervisor_tests;
-    use super::utils::PaddlerWorld;
-
-    #[tokio::test]
-    async fn test_all() {
-        run_supervisor_tests().await;
-        // run_agent_tests().await;
-        // run_balancer_tests().await;
-    }
-}
