+++ integration/supervisor-1.toml
@@ -1,0 +1,2 @@
+[config]
+supervisor-1 = '["--args","llama-server","-m","/usr/local/home/models/qwen2_500m.gguf","-np","4","--port","8080","--slots"]'
