+++ integration/tests/mod.rs
@@ -1,0 +1,16 @@
+pub mod utils;
+
+// #[cfg(test)]
+// mod tests {
+//     // use super::agent::tests::run_agent_tests;
+//     // use super::balancer::tests::run_balancer_tests;
+//     use super::supervisor::run_supervisor_tests;
+//     use super::utils::PaddlerWorld;
+
+//     #[tokio::test]
+//     async fn test_all() {
+//         run_supervisor_tests().await;
+//         // run_agent_tests().await;
+//         // run_balancer_tests().await;
+//     }
+// }
